use floats

+++ create_test/create_cache_optimal_bpf_program.py
@@ -11,7 +11,7 @@ class Type:
     FLOAT = "float"
     INT = "int"
 
-mat_type = Type.INT
+mat_type = Type.FLOAT
 
 
 def get_begin(mat_dim: int, rolled: bool):
